+++ R/tidyselect_to_list.R
@@ -24,8 +24,11 @@ tidyselect_to_list <- function(.data, x, .meta_data = NULL, input_type = NULL) {
   # check class of input -------------------------------------------------------
   # each element must be a formula or a named element of a list
   is_formula <- map_lgl(x, ~ class(.x) == "formula")
-  if (is.null(names(x))) has_name <- rep(FALSE, length(x))
-  else has_name <- as.logical(names(x) != "")
+  if (is.null(names(x))) {
+    has_name <- rep(FALSE, length(x))
+  } else {
+    has_name <- as.logical(names(x) != "")
+  }
 
   if (!all(is_formula | has_name)) {
     example_text <-
@@ -34,15 +37,25 @@ tidyselect_to_list <- function(.data, x, .meta_data = NULL, input_type = NULL) {
         "type" = paste("type = list(vars(age) ~ \"continuous\", all_integer() ~ \"categorical\")", collapse = "\n"),
         "label" = paste("label = list(vars(age) ~ \"Age, years\", vars(response) ~ \"Tumor Response\")", collapse = "\n"),
         "statistic" = paste("statistic = list(all_continuous() ~ \"{mean} ({sd})\", all_categorical() ~ \"{n} / {N} ({p}%)\")",
-                            "statistic = list(vars(age) ~ \"{median}\")", collapse = "\n"),
+          "statistic = list(vars(age) ~ \"{median}\")",
+          collapse = "\n"
+        ),
         "digits" = paste("digits = list(vars(age) ~ 2)",
-                         "digits = list(all_continuous() ~ 2)", collapse = "\n"),
+          "digits = list(all_continuous() ~ 2)",
+          collapse = "\n"
+        ),
         "value" = paste("value = list(vars(grade) ~ \"III\")",
-                        "value = list(all_logical() ~ FALSE)", collapse = "\n"),
+          "value = list(all_logical() ~ FALSE)",
+          collapse = "\n"
+        ),
         "test" = paste("test = list(all_continuous() ~ \"t.test\")",
-                       "test = list(vars(age) ~ \"kruskal.test\")", collapse = "\n"),
+          "test = list(vars(age) ~ \"kruskal.test\")",
+          collapse = "\n"
+        ),
         "mixed" = paste("label = list(vars(age) ~ \"Age, years\")",
-                        "statistic = list(all_continuous() ~ \"{mean} ({sd})\")", collapse = "\n")
+          "statistic = list(all_continuous() ~ \"{mean} ({sd})\")",
+          collapse = "\n"
+        )
       )
 
     stop(glue(
@@ -56,8 +69,10 @@ tidyselect_to_list <- function(.data, x, .meta_data = NULL, input_type = NULL) {
   # converting all inputs to named list ----------------------------------------
   named_list <-
     imap(
-      x, ~tidyselect_to_list_one(.data = .data, x = .x, x_name = .y,
-                                 .meta_data = .meta_data, input_type = input_type)
+      x, ~ tidyselect_to_list_one(
+        .data = .data, x = .x, x_name = .y,
+        .meta_data = .meta_data, input_type = input_type
+      )
     ) %>%
     flatten()